Added the beginnings of a database interface in C++ for the Hymnal.
DatabaseHelper.cpp will be the main C/C++ API to call into the database and functions there will return Hymn and Hymnal objects when required. 

Also, uncommented some lines of DatabasePopulator.py that seemed important to running it and commented out the line that specified a folder on Jake's computer.