Added support for unison staves.
DatabasePopulator.py now looks for treble staves that are at the end or
are followed by another treble staff, and marks them as "unison (0).

Added hymn 77, which come to think of it is probably not actually from
Hymnal: A Worship Book, but anyway it demonstrates that the script works.

Removed androidApp/bin from the repository. This is cleaner, and I'm pretty
sure nothing bad will happen.