Fixed a few more niggling errors in the display of a Hymn. 1. Hymns start displayed at the correct Zoom in Landscape mode. 2. Hymns are shown at the correct verse always (I hope) when rotating from portrait to landscape.
Added a method to hide the Navigation and Status bars when displaying a hymn to maximize available screen space. 
This happens on a Timer that fires Two seconds after the view is displayed. 

For now, tap the top left corner of the view to get the Navigation bar back.  I'm not sure what the best mechanic for this is. This mimics Read It Later's fullscreen for now.  Maybe it would be better for it to happen on a tap anywhere on the screen a la NY Times. 

Fixed an issue in the example where the Navigation Controller got confused because the same ViewController was being added to two separate Navigation controllers.